fix: Only show "not enough data" text when there isn't enough data. (#10604)
The previous implementation had a small bug where it would show "not
enough data" even if the graph was loading. This commit fixes that.

While working, I thought that maybe we should keep the current data
while we're fetching new data instead of adding a cover, but all the
other graphs use a cover when loading, so I've not made any changes to
that effect.