chore: scroll to the top of table when changing page
Makes it so that if you change the page in a paginated table, you'll scroll to the top of the table. Makes the experience more user friendly.

Will only scroll if the top of the table isn't in view already.

Finally: should we also scroll when you change the number of result rows per page?