chore: add Change Requests to sidebar and fix command menu icon
Adds the new change request overview to the sidebar and fixes the entry in the command bar to use the right title and icon.

Additionally, standardize the change requests icon. Instead of using a custom one from a svg, let's use the material icon (which is the same, just slightly flipped).

This standardization, makes the icon a tiny bit smaller in the new project modal, but it's the same size (20x20) as the other icons, so I don't think anyone's gonna notice or care, and I think it's better to use the basic material icon.