fix: don't clear "implicit" filters when updating table state outside quick filters
Fixes bug that would only occur if the first thing you do (when there's no url query params) on the page is to try to change the sort order or change the number of results per page.

In those cases, the table state would be replaced with only the new state from the sorting/page limit (and probably page).

In more specific terms: if you're on the page with no query params, then that means that you're seeing your open change requests.

But if you tried to change the sorting, say, then the "state" and "createdBy" filters would be cleared, and you would end up showing "all change requests ever".

The fix is to spread the implicit table state into the new state before updating the actual state, such that implicit filters become explicit when that happens.

## Implicit filters?

So why do we have implicit filters? Partly aesthetic, partly because that's how it works on github (github.com/pulls), and partly because that makes it easier to share with coworkers. You just need to go to the change requests page and copy the url. With no query params, they'll see their own results instead of yours.