fix: handle invalid permissions in role creation
The `create` and `update` role methods used to blindly accept any incoming permissions, but if the permissions don't exist in the database, then the database would throw, yielding a 500 error to the user.

To fix this, we can validate that all the permissions exist before we try to add the incoming permissions.

The http error only manifests in enterprise, but the fix requires modifying the access service. Therefore, I've added the tests to the access service too, such that if you break something, then you don't need to wait for it to propagate to enterprise.