fix: remove unusable button from navigation sidebar accordion
For some reason, there's two nested buttons here, where only the outer one is interactable. It doesn't matter much for mouse interactions, but for keyboard interactions, they're both focusable (which is strange) and the inner one doesn't do anything.

As such, this PR:
- removes it from the navigation tree (tabindex -1)
- changes the component to be a span instead of a button.

It keeps the role=undefined to override the default role="button" that MUI sets on it.