fix: sidebar overflow when displaying badges
Fixes a bug where the sidebar accordion would overflow if one of the children had a badge. This was visible primarily in pro instances up until now, but will also be an issue with the new "new" badge.

Instead of basing the width and min-width on the "mode", we now set it to `auto` regardless.

This means that:
- When there's no badges, the sidebar will be slightly narrower than it is today, because it doesn't need all that space.
- When there *are* badges, it will be slightly wider.

The previous change to these width values was in https://github.com/Unleash/unleash/pull/8831 and was apparently to make it "easier to use on medium screen sizes".

While this change makes it slightly wider when there are badges, I think that's a worthwhile tradeoff to:
- not having the accordion body be wider than the rest of the menu
- not having badges have to be smooshed up against the text

There's no max width set, so technically you could force it super wide by having very long menu items, but I think that's an acceptable tradeoff that we can make. I also don't think it's very likely that we'll run into that issue and not notice.

Further, on narrow screens, the sidebar is replaced by a modal anyway, so it's not likely to be much of an issue, I think.