fix: don't unselect selected path in configure menu when clicking it again
This fixes a bug where if you re-clicked the current path you were on in the configure sub-menu, it would unselect the item, even though if that's where you were still at.

The reason is that the onClick handler was pre-populated with the text `configure`. (However, the children were calling onClick with their own paths). I'm not really sure why this even worked in the first place.

Anyway, now the types have been changed and I've removed the application of the "configure" string as the active item.