chore: remove unused `route.route` property + rename function (#11039)
This removes the `route.route` property that was previously set within
the `mapRouteLink` function.

This property never existed on the `INavigationMenuItem` interface, so
the fact that it was assigned to was only allowed because the function
return type was implicit.

Further, this property seems to have never been used. There's some
references to the `route` property in the command bar code, but the
`route` property is never accessed there either.

As such, I think it makes sense to remove it. Additionally, because the
`mapRouteLink` function doesn't add any new properties, I thought it
made sense to rename it to `normalizeRoutePath`, as that is what it does
(removes the literal string `/*` from the route path).