add "new" badge to new entries and release management (#11038)
Adds a configurable "new" badge that can be used in the navigation
sidebar. It's configurable via a feature flag. It can show on either the
configure-menu or the primary menu (but not in the admin menu for now).
Only one item can be "new" at the same time.

## Feature flag config

Initially, I thought I was very clever when I added the configuration
through the feature flag, as it meant that we can easily switch the flag
on/off when we want to. But then I remembered that we also have
self-hosted users that won't get this. For now, I think that's fine,
because the flag will default to off anyway. We can, when we're ready,
set the default to `release management` or whatever we want, or we can
use a different implementation. We don't need to land that yet.

## Tests

There's no tests in this PR. However, I'll happily do a follow-up where
I add tests for the `filterMapRoutes` function. But ... also, I'm happy
to leave it 🤷🏼 I'll take input!

## Stack data

Stacked on top of
[#11039](https://github.com/Unleash/unleash/pull/11039) (but can be
untangled if necessary)

## Screenies

<img width="279" height="383" alt="image"
src="https://github.com/user-attachments/assets/dcb00c7a-985e-4156-a468-df18a79c092c"
/>

<img width="288" height="189" alt="image"
src="https://github.com/user-attachments/assets/ebd51ad2-fb34-4144-bf94-608ef08079e0"
/>

Is overridable with the feature flag, works for any routes in the main
menu.
<img width="233" height="147" alt="image"
src="https://github.com/user-attachments/assets/b1aec3c8-b2ba-42ff-8e49-739f6590da46"
/>