chore: don't use z-index hacking to hide mountain image
Removes the 'z-index: 100' from the footer and sets its position to `relative`. I presume that the only reason the Z-index was set explicitly was to cover the mountain background image. However, it's not necessary, and we can remove the arbitrary z index value as well as remove the explicit z-index setting on the image (0 is the default).

From what I can tell, this still works just fine with the app in various settings.