fix: notFlag logic in route (#11077)
https://linear.app/unleash/issue/2-4060/fix-route-notflag-logic

This fixes a regression introduced in
https://github.com/Unleash/unleash/pull/11063, where a `notFlag` would
always take precedence if present in the route configuration, even when
the flag was disabled.

What we actually want is for `notFlag` to take precedence **only when it
is set and its flag evaluates to true**. In other words: if a route
defines a `notFlag` and that feature flag is enabled, the route should
be filtered out. Otherwise, the remaining checks should still apply.