chore: add new NewInUnleashItems file (#11086)
Adds a NewInUnleashItems file. This file will supplant the list of items
coded into the existing `NewInUnleash.tsx`. Because we can use links
instead of onClick navigation handlers for app navigation, we also don't
need it to be stateful. If there is a need for this later, we can add it
in then.

The file also exports a type for the items.

Also, removes the previous items (lifecycle 2.0, and signals and
actions) from the list (or rather: doesn't port them over) because
they're out of date now anyway.