chore: Add new `NewInUnleash` component.
This is primarily a wrapper component that finds which item to show for "new in unleash" and then passes that wrapper to the toast.

Also extracts the logic that decides whether an items should be displayed and adds tests for it. This logic was untested in the legacy implementation. (It's a fairly low-stakes check, though, so happy to take the tests out if you don't like 'em).