chore: add a delayed evaluation of UI flags
Adds a new hook for UI flag evaluation that returns an evaluation function.

This solves a frequent pain point where we can't check if a flag is enabled conditionally (e.g. if we don't know if we have a flag yet) because it's implemented as a hook.

The new implementation uses the same evaluation as the old one, but returns the eval function after calling the hook, thereby circumventing the issue and allowing you to eval a flag anywhere in the function body.