Update `NewInUnleashDialog` to accept app links as well as `onCheckItOut` (#11087)
Updates the `NewInUnleashDialog` to accept elements that have an
`appLink` prop as well as elements with an `onCheckItOut` prop, allowing
for the changes in #11086 to be displayed the same as the old items. For
items with an `appLink` prop, the "check it out" button will be rendered
as a link instead of a button.

The dialog otherwise looks and acts exactly the same.

Builds on https://github.com/Unleash/unleash/pull/11086