add new new in unleash toast (#11066)
Adds a "new in unleash" toast (*snackbar* in MUI parlance) to the app.
The component is not hooked up to anything, so it will not show anywhere
in the app just yet.

Because we want this to show regardless of whether the sidebar is open
or not and whether the user is on mobile or not, the new components need
to be mounted outside of the sidebar hierarchy. Because they're on top
of everything, I've used the main layout in dev, but that can be
adjusted if necessary. Nothing is mounted there at the moment, though.

The toast acts like a "card" in many ways, where the entire thing is
clickable and will open up the modal. There is no implementation for
modal-less items for now because the current thing we're highlighting
(release plans) uses a modal. We can add support for modal-less items
later.

On component mount, there is a 500 ms timeout before the toast slides
in.

Stacked on top of https://github.com/Unleash/unleash/pull/11087

---

Close-up of toast + with focus styles:

<img width="330" height="119" alt="image"
src="https://github.com/user-attachments/assets/0997d772-3174-48dc-b0c0-affedd2c1815"
/>

<img width="315" height="118" alt="image"
src="https://github.com/user-attachments/assets/ae643d30-1434-4879-a294-24b5375c5c37"
/>

The toast goes behind the overlay when the modal shows (snackbars have a
higher z-index than modals do in MUI by default):

<img width="1490" height="1253" alt="image"
src="https://github.com/user-attachments/assets/97ee2216-8991-47f6-bf32-764bbe230e74"
/>

## Up next

Add a wrapper component that checks which news item (if any) to show and
that will be mounted in the main layout.

+ Add a flag to control which version of the new in Unleash we're
showing.


## Dark mode

The colors aren't correct in dark mode. Or rather, the `common.black`
theme color isn't very black. Because it's not wired up anywhere, I'm
happy to leave it like this for now while we wait for UX to get back to
us with what they wanna do.

<img width="322" height="125" alt="image"
src="https://github.com/user-attachments/assets/461b1c1e-3886-41ac-abb0-806a06f9b554"
/>

## Hover colors

There's also no hover indicator at the moment. I'm waiting for input
from UX for that too, but that's a small thing to add.