feat: add project-specific edit and create modal forms (#11127)
Allows you to create, edit, and delete project context fields from the
project context field page.

Follows the same way of doing dynamic routing on the buttons as we did
for segments (checking if projectId exists, and if so, route there).

The current impl blindly inserts project if the edit/create form is open
under a project path, but we probably want to parameterize this later
(or before merging), so as to make it easier to control.