fix(chore):  update the UI to use project-based context endpoints when on a project endpoint (#11160)
Reverts #11156 which was itself a revert of #11153 and then applies a
fix to two breakages
- one in the strategy configuration overview
- one in the context list

## Strategy edit context bug:
Fixed in ef6f6542bca13ab9e74993031d3af270605f2661

No context options were available in the strategy edit modal:

<img width="854" height="904" alt="image"
src="https://github.com/user-attachments/assets/6c8be391-294d-4b31-85ce-492db53c1b76"
/>

The fix itself was removing a double call to `formatApiPath`, which
would double up the base path if it's not an empty string, e.g.
`/sandbox/sandbox/api[...]` instead of `/sandbox/api[...]`.

When the base path was doubled up, Unleash would attempt to return the
base app (HTML) which obviously can't be parsed into anything useful.

## Context list bug
Fixed in c410daf183795b0090638f4178ecae9b22688e0a

Edit links would point to `context/edit/undefined` and editing via the
pen also caused some weird issues.

The main cause was that the `name` variable was shadowed and I didn't
realize. So when the edit string was created, it wouldn't be the context
field's name, but instead the name of the context field to delete (for
which there is a weird, stateful solution -- I remember there was a bug
that we fixed with it).

This has been fixed by moving the edit url variable declaration further
down, such that the `name` variable is in scope. additionally, I've
renamed the generic `[name, setName]` variables to
`[contextFieldToDelete, setContextFieldToDelete]` to avoid shadowing and
to make it clearer what the name is being set for.

## Scenarios tested 
Against sandbox instance with a base path (using a viewer user with
project update and context field crud permissions):

Flag off:
- [x] Create / edit / delete context field in global menu
- [x] Context fields with projects (as created with the flag on) are
show in global list when flag is off
- [x] Context fields with projects can be edited in the global list (but
are still project-based if the flag goes back on)
- [x] Context fields with projects can be deleted in the global list.
- [x] Select and edit context fields on a strategy

Flag on:
- [x] Create / edit / delete context field in global menu
- [x] Create / edit / delete context field in project menu 
- [x] Context fields with projects are not shown in global menu
- [x] Context fields not belonging to "this" project are not shown in
project-scoped menu
- [x] Select and edit (global) context fields on a strategy

## Known limitations:

Using project-scoped context fields for strategies is not yet supported.